fetching images w/ on(load)
images weren't completely loading before being displayed, so they would
snap into place some times, but not always. using the on(load) function,
the images are loaded before any of the animations are performed

there's still some work left to be done here. right now the way the URL
is crafted is convoluted and wasteful. will tighten it up